requirements.txt
# Core dependencies
ccxt==4.1.77
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
python-dotenv==1.0.0
requests==2.31.0

# Data visualization
plotly==5.17.0
dash==2.16.1
dash-bootstrap-components==1.5.0
kaleido==0.2.1
seaborn==0.12.2
matplotlib==3.8.2

# Machine learning & Deep learning
tensorflow==2.15.0
xgboost>=1.6.0
torch>=1.12.0
transformers>=4.20.0

# Portfolio optimization
cvxpy>=1.2.0
scipy>=1.9.0

# Technical analysis
ta==0.10.2

# Sentiment analysis & NLP
tweepy==4.14.0
textblob==0.17.1
nltk==3.8.1
vaderSentiment==3.3.2
praw>=7.6.0

# Web scraping & APIs
beautifulsoup4==4.12.2
lxml==4.9.3
feedparser>=6.0.0

# Communication
python-telegram-bot==20.7

# Web framework
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
pyjwt==2.8.0
werkzeug==3.0.1

# Database & Caching
redis==5.0.1
celery==5.3.4

# Additional utilities
schedule==1.2.0
apscheduler==3.10.4
yfinance==0.2.28
websocket-client==1.7.0
python-binance==1.0.19
alpaca-trade-api==3.1.1
wordcloud==1.9.2
gspread==5.12.4
oauth2client==4.1.3

# Advanced ML libraries
lightgbm>=3.3.0
catboost>=1.0.0
statsmodels>=0.13.0

# Risk management
pyfolio>=0.9.0

# Additional utilities
colorama>=0.4.0
tqdm>=4.64.0
aiohttp>=3.8.0 